% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genre_mean_critics.R
\name{calc_genre_mean_critics}
\alias{calc_genre_mean_critics}
\title{Video Game Mean Critic Score Calculator}
\usage{
calc_genre_mean_critics(.data)
}
\arguments{
\item{.data}{A data frame containing a column named `critic_score`}
}
\value{
Returns a tibble with a column called `Genre` and another called `Mean Critic Score`
followed by a warning message that all NA values were dropped, including emojis and red text.
}
\description{
This function calculates the mean critic score and presents it in a formatted tible,
including a custom message near the end.
}
\details{
This function takes in a data frame that contains a column named `critic_score`. After
doing necessary data wrangling using tidyr and dplyr, the function will calculate the mean
critic score for each video game genre. At the end, "Warning: NA values were dropped" is written in
red and is flanked by two "anger" emojis.
}
\examples{
## Not run:
## vg_data <- read.csv("vgchartz-2024.csv")
## calc_genre_mean_critics(vg_data)
## End(Not run)

}
