% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time_diff.R
\name{calc_time_diff}
\alias{calc_time_diff}
\title{Video Game Release Date & Update Date Time Difference Calculator}
\usage{
calc_time_diff(.data, x)
}
\arguments{
\item{.data}{A data frame containing a column named `release_date` and a
column named `last_update`}

\item{x}{The data frame's row number; this has to be an integer value}
}
\value{
Returns a sentence that says "Time difference of ___ weeks";
where ___ represents the calculated time difference
}
\description{
This function calculates the time difference between the video game’s release date
and the date it was last updated.
}
\details{
This function takes in a data frame containing a column named `release_date` and a
column named `last_update` as well as an integer that represents the row number of interest.
After making sure all needed values are in correct format, the function checks to see if
the `release_date` and/or `last_update` columns have NA values. If so, custom errors are returned.
Assuming that both columns don't have NA values, the time difference between each is calculated.
}
\examples{
name <- c("Just Dance 4", "Call of Duty: Black Ops 3")
dates1 <- c("2018-01-31", "2018-01-14")
dates2 <- c("2012-10-09", "2015-11-06")
df <- data.frame(video_game_name = name, last_update = dates1, release_date = dates2)

calc_time_diff(df, 1)

}
