% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sales.R
\name{map_sales}
\alias{map_sales}
\title{Map global game sales}
\usage{
map_sales(game_title)
}
\arguments{
\item{x}{The dataframe 'vg_long' stored in data file, function includes a merge with geospatial
data}
}
\value{
Prints a map using the r package leaflet and the background map CartoDB.Positron.
}
\description{
This function creates a leaflet map of global video game sales for individual games.
Categories are displayed by appearance in the data: north american sales, japanese sales, european
and african sales, and all other countries.
}
\details{
This function takes the argument of a video game name. Ensure that the spelling of the
matches how it is spelled in the data. The shapefile for this data was accessed via rnaturalearth
data, containing polygons for each country in the world. This function combines those polygons into
lists, generating a column of multiple polygons outlining which countries are included in which sales
category.
}
\examples{
# map_sales("Grand Theft Auto V")
# map_sales("Red Dead Redemption 2")

}
