% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_console.R
\name{print_console}
\alias{print_console}
\title{Video Game Console Printer}
\usage{
print_console(.data, x)
}
\arguments{
\item{.data}{A data frame containing a column named `console` representing gaming consoles}

\item{x}{An integer value that represents the number of rows that will be printed.}
}
\value{
Returns a tibble with a column that says `Console Type` and another that says
`Number of Video Games`. The user chooses how many rows of the tibble will be printed.
}
\description{
This function prints the number of video games available for each gaming console.
}
\details{
This function takes in a data frame containing a column named `console` as well as an
integer that represents the numbers of rows that will be printed. After data wrangling the table
with dplyr functions, the function checks if the index value is within valid range. If not,
a custom error will be printed. If so, then a tibble with the desired number of rows will be printed.
}
\examples{
## Not run:
## vg_data <- read.csv("vgchartz-2024.csv")
## print_console(vg_data, 10)
## End(Not run)

}
