% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_installments.R
\name{print_installments}
\alias{print_installments}
\title{Print video game installments}
\usage{
print_installments(game_title)
}
\arguments{
\item{x}{The dataframe 'vgchartz_2024' stored in data file, function includes filtering the data
to unique observations only (no repeated titles)}
}
\value{
Prints a statement saying how many installments of a franchise exist in the data
}
\description{
This function prints the number of installments for a given video game title.
}
\details{
This function takes the argument of a video game franchise name. It searches through
all video games in the dataset for a matching pattern of characters. For instance, if "Red Dead"
is used as the argument, all video games with "Read Dead" in the title will be returned. The
function is not case sensitive.
}
\examples{
# print_installments("Red Dead")
### There were 9 installments found for games with Red Dead in the title
# print_installments("Red Dead Redemption II")
### There were 1 installments found for games with Red Dead in the title

}
