% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_developers.R
\name{graph_developers}
\alias{graph_developers}
\title{Graph global developer sales}
\usage{
graph_developers(dev_name)
}
\arguments{
\item{The}{name of the developer}
}
\value{
Prints a plot using the R package ggplot2 and a Brewer color palette.
}
\description{
This function creates a column plot of global game sales for specific developers
by region.
}
\details{
This function takes the argument of a developer name. Ensure that the spelling of the
matches how it is spelled in the data. The function will return a column plot showing the game
sales for the 4 regions in our data
}
\examples{
# graph_developers("Konami")
# graph_developers("Naughty Dog"))

}
